feat: add deployment URL
This commit, which has hashicorp/waypoint-plugin-sdk#27 as a pre-requisite,
adds support for Deployment URLs.

With this PR, it is now possible to show in the CLI the deployment URLs
provided by the plugins, without the need of Horizon (URL Service).

Warning: The PR contains a go.mod entry that points to our fork since
the PR in https://github.com/hashicorp/waypoint-plugin-sdk hasn't been
merged yet.